Fix #88 move io classes to their own module
