+++ tests/examples/testWorkflowExamples.cpp
@@ -1,11 +1,11 @@
 
 #include <catch2/catch_test_macros.hpp>
 
-#include "BaseIO.hpp"
+#include "io/BaseIO.hpp"
 #include "Channel.hpp"
 #include "Types.hpp"
 #include "Utils.hpp"
-#include "hdf5/HDF5IO.hpp"
+#include "io/hdf5/HDF5IO.hpp"
 #include "nwb/NWBFile.hpp"
 #include "nwb/RecordingContainers.hpp"
 #include "nwb/file/ElectrodeTable.hpp"
