+++ tests/examples/test_example.cpp
@@ -1,7 +1,7 @@
 // [example_all]
 #include <catch2/catch_test_macros.hpp>
 
-#include "hdf5/HDF5IO.hpp"
+#include "io/hdf5/HDF5IO.hpp"
 #include "testUtils.hpp"
 
 TEST_CASE("SimpleExamples", "[hdf5io]")
